refactor(web/supabase): update type imports and restructure types_db file
- Changed import paths for types in chat.ts and client.ts to reflect the new types_db location.
- Created a new types_db.ts file to define database types, enhancing type safety and organization within the project.
- Updated the generate-types script in package.json to output the types to the new directory.
