+++ apps/web/package.json
@@ -49,7 +49,7 @@
   "scripts": {
     "build": "next build",
     "dev": "next dev --turbopack",
-    "generate-types": "npx supabase gen types typescript --project-id hgkmrremnwapjdbebvqo --schema public > types_db.ts",
+    "generate-types": "npx supabase gen types typescript --project-id hgkmrremnwapjdbebvqo --schema public >./types/types_db.ts",
     "lint": "next lint",
     "start": "next start"
   },
